requirements.txt
google-generativeai>=0.7.0
google-api-python-client>=2.134.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
python-dotenv>=1.0.1
tenacity>=9.0.0
typer>=0.12.3
rich>=13.7.1
dataclasses-json>=0.6.7
pytest>=8.2.0
pytest-mock>=3.14.0
freezegun>=1.5.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
